fix(issue-1852): renaming
