+++ packages/react/src/patterns/blocks/FeatureCardBlockLarge/__stories__/FeatureCardBlockLarge.stories.js
@@ -7,12 +7,12 @@
 import './index.scss';
 import { object, text, withKnobs } from '@storybook/addon-knobs';
 import { ArrowRight20 } from '@carbon/icons-react';
-import FeatureCardLarge from '../FeatureCardLarge';
+import FeatureCardBlockLarge from '../FeatureCardBlockLarge';
 import React from 'react';
 import readme from '../README.md';
 import { storiesOf } from '@storybook/react';
 
-storiesOf('Patterns (Blocks)|FeatureCardLarge', module)
+storiesOf('Patterns (Blocks)|FeatureCardBlockLarge', module)
   .addDecorator(withKnobs)
   .addParameters({
     readme: {
@@ -44,7 +44,7 @@ storiesOf('Patterns (Blocks)|FeatureCardLarge', module)
       <div className="bx--grid">
         <div className="bx--row">
           <div className="bx--col-sm-4 bx--col-lg-12 bx--offset-lg-2">
-            <FeatureCardLarge
+            <FeatureCardBlockLarge
               heading={heading}
               card={{
                 heading: cardheading,
