+++ packages/react/src/patterns/blocks/FeatureCardBlockLarge/__stories__/index.scss
@@ -1,0 +1,1 @@
+@import '../../../../../../styles/scss/patterns/blocks/feature-card-block-large/feature-card-block-large';
