--- packages/react/src/patterns/blocks/FeatureCardLarge/__stories__/index.scss
@@ -1,1 +1,0 @@
-@import '../../../../../../styles/scss/patterns/blocks/feature-card-large/feature-card-large';
